+++ microlens-platform/src/Lens/Micro/Platform.hs
@@ -37,6 +37,7 @@ module Lens.Micro.Platform
   module Lens.Micro.GHC,
   module Lens.Micro.Mtl,
   module Lens.Micro.TH,
+  packed, unpacked,
 )
 where
 
@@ -45,6 +46,7 @@ import Lens.Micro.Internal
 import Lens.Micro.GHC
 import Lens.Micro.Mtl
 import Lens.Micro.TH
+import Lens.Micro.Platform.Internal
 
 import Data.Hashable
 import Data.Int